requirements.api.txt
aio-pika
fastapi[standard]
python-dotenv
